lcs is hard

+++ template/mafs/macros.cpp
@@ -3,7 +3,7 @@
 #define add(x, y) (((x) % MOD + (y) % MOD) % MOD + MOD) % MOD
 #define sub(x, y) add((x), -(y))
 #define mul(x, y) (((x) % MOD * (y) % MOD) % MOD + MOD) % MOD
-#define div(x, y) long(x) * inv((y)) % MOD
+#define div(x, y) x * inv((y)) % MOD
 
 using namespace std;
